splitting up functions from coding_effect to make refactoring easier
